refactor: PokemonCard changed to functional components
