+++ src/components/PokemonCard/PokemonCard.module.css
@@ -17,6 +17,10 @@
   margin-bottom: 10px;
 }
 
+.name::first-letter {
+  text-transform: uppercase;
+}
+
 .img {
   width: 150px;
   height: 150px;
