--- src/interface/PokemonCardProps.ts
@@ -1,5 +1,0 @@
-export interface PokemonCardProps {
-  name: string;
-  img?: string;
-  id: number;
-}
